build: change bundler to tsdown and upgrade to ts6
