+++ pnpm-workspace.yaml
@@ -1,27 +1,27 @@
 catalogs:
   actions:
-    '@actions/core': ^3.0.0
-    '@actions/github': ^9.0.0
+    '@actions/core': ^3.0.1
+    '@actions/github': ^9.1.1
   build:
-    terser: ^5.46.0
-    tsup: ^8.5.1
-    typescript: ^5.9.3
+    terser: ^5.46.1
+    tsdown: ^0.21.9
+    typescript: ^6.0.3
   ci:
-    '@commitlint/cli': ^20.3.1
-    '@commitlint/config-conventional': ^20.3.1
-    '@favware/cliff-jumper': ^6.0.0
+    '@commitlint/cli': ^20.5.0
+    '@commitlint/config-conventional': ^20.5.0
+    '@favware/cliff-jumper': ^6.1.0
     husky: ^9.1.7
-    lint-staged: ^16.2.7
+    lint-staged: ^16.4.0
   core:
     '@npm/types': ^2.1.0
-    '@types/bun': ^1.3.6
-    commander: ^14.0.2
+    '@types/bun': ^1.3.12
+    commander: ^14.0.3
   lint:
-    '@nanoforge-dev/utils-eslint-config': ^1.0.1
-    '@nanoforge-dev/utils-prettier-config': ^1.0.1
+    '@nanoforge-dev/utils-eslint-config': ^1.0.2
+    '@nanoforge-dev/utils-prettier-config': ^1.0.2
     '@trivago/prettier-plugin-sort-imports': ^6.0.2
-    eslint: ^9.39.2
-    prettier: ^3.8.0
+    eslint: ^10.2.1
+    prettier: ^3.8.3
 
 onlyBuiltDependencies:
   - esbuild
