+++ tsconfig.json
@@ -35,14 +35,15 @@
 
     // Interop Constraints
     "allowSyntheticDefaultImports": true,
-    "esModuleInterop": false,
     "forceConsistentCasingInFileNames": true,
     "isolatedModules": true,
 
     // Language and Environment
-    "experimentalDecorators": true,
-    "lib": ["ESNext", "esnext.disposable", "DOM"],
+    "rootDir": ".",
+    "lib": ["ESNext", "esnext.disposable"],
+    "types": ["bun"],
     "target": "ESNext",
+    "experimentalDecorators": true,
     "useDefineForClassFields": true
   },
   "include": ["src/**/*.ts"],
